Solve sensor readin reply sometimes truncated
After prolonged use, the final part of the json sensor reading
(a long text string) was sometimes missing. Seems to be due to
memory fragmentation and / or the firmware doing background tasks.

Now we permanently reserve the memory needed at program start,
and have been running for months without seeing the issue once.